Refactor integer types
Major type refactoring to address root causes of type conversion warnings
instead of masking them with casts. This improves type safety, enables
better compiler optimizations, and makes the codebase more maintainable.

Core changes:
- Changed xmalloc/xrealloc/xrecalloc to use size_t for memory sizes
  (the proper type for memory allocation)
- Updated all memory tracking variables (memsize, memused, maxmemsize)
  to size_t for consistency
- Changed game value[][] array from int to uint16_t to match network
  protocol (load_u16) and game semantics (stats can't be negative)
- Changed rage/endurance from int to uint16_t for same reasons
- Fixed sdl_image xres/yres from int16_t to uint16_t (image dimensions
  are never negative)
- Refactored map selection indices (mapsel, itmsel, chrsel) to size_t
  for semantic correctness as array indices
- Updated related functions to return uint16_t instead of int/unsigned int
  (tactics2melee, tactics2immune, tactics2spell, has_action_skill)

Philosophy applied:
- Use smallest appropriate unsigned type (uint16_t vs unsigned int)
- Fix types at declaration rather than adding casts
- Use load_u16() instead of casting load_i16() for protocol data
- Only cast at legitimate type boundaries or for arithmetic coercion

This removes hundreds of implicit conversion warnings and significantly
reduces the number of casts in the codebase.

+++ src/amod/amod.h
@@ -60,9 +60,9 @@ DLL_IMPORT int doty(int didx);
 DLL_IMPORT int butx(int bidx);
 DLL_IMPORT int buty(int bidx);
 // gui helperls
-DLL_IMPORT unsigned int get_near_ground(int x, int y);
-DLL_IMPORT unsigned int get_near_item(int x, int y, unsigned int flag, unsigned int looksize);
-DLL_IMPORT unsigned int get_near_char(int x, int y, unsigned int looksize);
+DLL_IMPORT size_t get_near_ground(int x, int y);
+DLL_IMPORT size_t get_near_item(int x, int y, unsigned int flag, unsigned int looksize);
+DLL_IMPORT size_t get_near_char(int x, int y, unsigned int looksize);
 DLL_IMPORT unsigned int mapmn(unsigned int x, unsigned int y);
 // misc
 DLL_IMPORT void set_teleport(int idx, int x, int y);
