+++ src/astonia.h
@@ -161,9 +161,9 @@ DLL_EXPORT int fail(const char *format, ...) __attribute__((format(printf, 1, 2)
 DLL_EXPORT void paranoia(const char *format, ...) __attribute__((format(printf, 1, 2)));
 void display_messagebox(char *title, char *text);
 
-void *xmalloc(int size, int ID);
+void *xmalloc(size_t size, int ID);
 void *xrealloc(void *ptr, size_t size, int ID);
-void *xrecalloc(void *ptr, int size, int ID);
+void *xrecalloc(void *ptr, size_t size, int ID);
 void xfree(void *ptr);
 char *xstrdup(const char *src, int ID);
